Exit with status code after kernel_main/panic
This is a prereq for cargo test. This allows us to flag to the caller if
our kernel ran as expected. For now on panic exit 1, and on normal
completion exit 0.

qemu_wrapper.sh is required as the implementation of the exit() call is
(code << 1) | 1, we need to undo this to get us back into normal
expected exit code ranges