Hide inb/outb commands behind port type
A problem with the existing implementation was that there was no way to
guarantee drivers were not clobbering eachother. We can solve this with
some type system tricks.

* Remove inb/outb commands that were previously used. Hide them behind a
  Port type
* Add PortManager that is the only way to get a Port
* PortManager ensures that each port is only handed out once

With this scheme we know that multiple drivers aren't using the same
items.

Re-write the serial driver to use explicit port names instead of
inb/outb with port + offset.
  * With the serial driver now having explicitly named ports, and ports
    can fail to be acquired, add a new enum error type and pull in
    thiserror-no-std to create it

Move static variables for stdout to io module, and initialize with
init_stdio function. We want stdio to be initialized as soon as
possible.

Move stdio init after alloc init. This is an unfortunate side effect of
using a HashMap in the PortManager implementation

Add new static port for exit so io::exit(u8) still works nicely