Initial loading of global descriptor table
This is a pre-requisite to getting interrupts working. The multiboot
spec specifies that a gdt is loaded by multiboot before loading, however
the value in the gdtr register may not be valid. Because of this we need
to re-initialize with our own gdt so that on interrupt we can load a
segment that we know is valid.

Add a GDT module, which creates a table in static memory, and loads it
into the gdtr register. This requires a bunch of manual bit twiddling,
so introduce some helpers for getting/setting bits as well.

Print the gdt before and after for confirmation that it's loaded
correctly