Interrupt implementation
Basic implementation of cpu interrupt handling

Interrupts are handled through an interrupt table. This is loaded with
an idtr command into the CPU. The CPU on interrupt will check this table
and load a gate descriptor from it to determine what code to run.
Generate an interrupt table and stub in a couple example interrupt
handlers

* Generate an interrupt table with gate descriptors
* Add handlers for a couple interrupts and print/abort
* Disable all external interrupts by masking them in the PIC
* Bonus fix for out of bounds access in TerminalWriter
* Run an interrupt as an example