Logger implementation
Implement a logger that's safe to use from the context of an interrupt
handler. In order to support this we want to avoid waiting long amounts
of time for the print outputs to finish. What we can do instead is push
logs into a buffer that is then periodically flushed from kernel main.

Add a interrupt guard class, as well as a static circular buffer
(instead of using VecDeque which could cause an allocation from an
interrupt that we don't want if we can avoid it). Use these to build a
logger that dumps buffered logs on service. Currently we just call
service once before quitting, but in the future this will be called as
part of our event loop

We mistakenly implemented a spinlock during implementation. Leave it in
for now as it will certainly be useful in the future

Now that we have a logger, move some previous print spam into debug logs