Revert to synchronized println
Initialization debugging proved to be difficult when we needed async to
be up and running in order to println. Given that we essentially never
wait for the serial buffer to be empty, revert back to the busy loop
style of waiting for the output buffer.

Keep print registration function in IO though, that was a nice way to
let the kernel own construction of the necessary drivers.

Kernel init adjusted to get printing as early as possible. While in
here, it was observed that Kernel::init() ran no async code, re-arrange
code to...

* Get println working ASAP
* Remove async on kernel::init