Add read/write to rtl8139 driver
Implement actual use of the rtl8139 to send/receive packets. This was
implemented with help from the osdev wiki, the rtl8139 datasheet, and
the rtl8139 programming guide.

At a high level, we have to
* Initialize the card
* Provide it a buffer to write incoming packets to
* When an incoming packet is read, update a circular buffer pointer to
  tell the card it is ok to write there
* On write we provide a contiguous buffer of physical memory to the
  device and ask it to write it to the network
* On subsequent writes we rotate through the 4 transmission registers

Write/read have been modeled with async functions. These poll the status
registers for outgoing writes, and polls the circular buffer pointers
for incoming reads

Supporting changes:
* Un-publicize register read/writes from Pci interface, external users
  should be using higher level commands
* Prevent double allocation of PCI device
* Fold sleep demo into async send/receive with loopback demo
* Async mutex added to allow simultaneous read/write