Reduce poll frequency
The existing logger implementation used a sleep timer to trigger
printing of logs. This is easy to fix with a future that polls to see if
there are any logs available. Now if we log iterations of future poll we
see a lot less wakeups

Slight re-order of executor loop to ensure that if a future tries to
wakeup the executor without an interrupt being run a new loop through
the executor will be triggered