Initial TCP implementation
Enough of TCP is implemented to support a simple echo server. This is
very much an unfinished implementation though. A lot of options are not
respected, and due to lack of retransmission this is definitely not a
reliable implementation.

That being said, we have a good starting point, and there's enough
"correct" code already in here to start being useful. For now we will
commit the unfinished implementation, and fix problems as we move
forward

* Add a TCP subsystem to the top level Kernel
* TCP class works as follows
  * It is unaware of ipv4/ethernet layer
  * When a tcp frame comes in, it parses and exports a tcp response if
    necessary
  * Users of tcp can create listeners/connections that can write out
    data to the other end of the connection
  * Tcp::service() needs to be called in a loop. This is where we handle
    things like outgoing messages, retransmission etc.
  * Tcp::service() will return a tcp frame that needs to be put on the
    network card
* EthernetFrame functions made public, they are needed to send out our
  tcp responses