Replace InterruptGuarded<T> in logger with lock free queue
The previous synchronization method for pushing logs to the global
logger was to use an interrupt guard. This prevented deadlocks that
would be caused from using a mutex in an interrupt, and prevented
multiple readers/writers as we knew nothing could preempt us while we
were modifying our internal data.

This makes the assumption that there is only one therad of execution. As
we work towards multi-core support, this is no longer a valid assumption

Add a lock free queue that can be used from an interrupt. This will
never block on push, but may need to retry a couple times in order to
find the correct index. This makes it safe to use in an interrupt
context, and is safe for multi-threaded code.

Use the lock free queue on the logger as a way to ensure that it is
working correctly

Lock free queue is split into a sender/receiver. This makes it easier to
ensure that there is a single reader of the queue (multiple writers are
allowed), which simplifies the implementation.

Additional changes:
* Always allocate in the log macro. The Cow<'static, str> optimization
  did not feel worth the duplication/complexity to me anymore
* Initialization of Logger is now done through an UnsafeCell<Option>. We
  could not initialize with a const fn anymore as internal types (e.g.
  atomics) were no longer const fns.
* A few extra clippy fixes
* Remove CircularArray. This has effectively been replaced with the lock
  free queue