Support writing to serial from an interrupt without crashing
If we printed from within an interrupt, we may be in a state where the
RefCell for the serial writer was already borrowed. Now we could argue
that we shouldn't be printing from an interrupt handler, but in some
cases it's important for debugging crashes.

Make the IoRange an UnsafeCell. This is technically a violation of
rust's mutability guarantees, but given that the implementation of
IoRange::write is almost entirely in asm I do not believe this will be a
problem.