Support sleep from multiple threads
The previous WakeupList assumed that an InterruptGuarded<T> would make
it safe to use from multiple contexts. This is true in the case where we
only have a single thread that can be preempted by interrupts. As we
move towards multi-core we now need to be able to request sleeps from
multiple threads

Replace the wakeup list with 3 components

* A mutex locked queue of sleep requests that can be posted to from
  multiple threads
* A servicer that migrates requests from all threads to the interrupt
  handler's context
* The end used from the interrupt

We need to use a mutex for multiple threads, but cannot lock a mutex
from the interrupt handler without a risk of deadlock. This scheme
creates a guarantee from the type system that only a single thread can
modify the wakeup list, and that an interrupt will not occur while this
is happening.

An alternative would be to try to use the lock free queue, however this
uses a pre-allocated bounded array. It's conceivable that we would issue
enough sleeps rapidly and run out of room in our queue. Using a
Mutex<VecDeque> avoids this and is safe enough.

We could also maybe argue that we could track the list in the interrupt
handler itself, however managing memory allocations from within the
interrupt handler seems like it could cause problems if it was too slow.

Tracking of wakeups now has to be a BTreeSet instead of a HashSet, as
sleeps could come in after the sleep duration has completed. A BTreeSet
gives us an ordered list of times that we can prune when they have
passed