Remove global wakeup_executor function
The existing wakeup of the future executor assumed that there was a
single global executor used for all futures. This was fine when we could
only run the executor once on a single core, but as we work towards
multi-core support we will want multiple executors running in parallel.
We need to know which executor to wake up in this case.

Remove the global wakeup_executor() function and propagate wakers to the
interrupt handlers that need them

* Add an "AtomicCell" type. This is a leaked AtomicPtr that should
  handle freeing when it is swapped out
* Use the AtomicCell to have a shared reference to a waker that is
  overwritten on each future poll
* For interrupts that have multiple futures waiting on them, introduce a
  service() function that propagates a single wakeup to all the
  individual futures. This scheme is chosen to avoid doing too much work
  from within the interrupt handler
  * On the rtl8139, this requires some registration/de-registration of
    futures that want to be woken. This is done with a spinlock +
    hashmap from future id -> waker. Spinlock was chosen because the
    async mutex type cannot be used from within the drop implementation
  * Note that the service future in Rtl8139::service() does not actually
    check for anything. This is because the dependent futures will check
    for themselves
* Bonus changes for spinlock type (since this is it's first real use)
  using my latest understanding of atomics