Fix high CPU usage on secondary cores
APs did not have interrupts enabled, and therefore had no way to wake up
the CPU after a hlt. This is our main mechanism of lowering CPU usage on
thread 1, so this ends up being quite a problem.

Enable the APIC on the secondary cores. Note that this is only the LOCAL
APIC, not the I/O APIC. The old (maybe emulated) PIC is still
dispatching external interrupts only to the BSP.

Send IPIs from BSP to APs when pushing a fn to the execution queue. This
will kick them out of the halted state

Note that some shortcuts have been taken in the interest of getting this
merged quickly enough to work on something new today

* Load interrupt descriptor table on APs
  * Split out load_idt function that stashes the descriptor table in
    static memory. On subsequent calls just re-load the previously
    generated table
* Add support for APICs in the generic interrupt handler (using a quick
  hack of comparing cpuid() to BSP id)
* Enable APICs + interrupts on APs
* hlt when no work to do on APs