PCI device instantiation improvements
Previously we were iterating all of PCI bus space for every PCI device
to find the specified ID. If we have more than one PCI device driver,
this does not make sense. Instead we should iterate the PCI bus space
once, and construct the driver we care about for each ID

We were also missing device function iteration. Some devices are
multi-function, and if we want to use multiple functions from those
devices, we need to find them. E.g. the UHCI device given when we use
the -usb flag in qemu is a secondary function of one of the default PCI
cards

While we're here, fix the device iteration. Previously it was iterating
from 0..=255, but device numbers are only valid up to 31